try to make nom installable
This is... wrong. It installs rakudo's pbcs into the general parrot install
location, because I can't figure out where the pbcs are actually loaded, and
how. On the plus side it gives us an idea what files need to be installed.

It gets us so far that we can actually run the perl6 binary from outside the
installation path, but using an Int literal results in "Could not locate
compile-time value for symbol Int"